Use usize for specifying limits
A bunch of the Alpaca endpoints accept some form of limit parameters.
Right now we specify some of them as u64 and others as usize. None of
these limits is really running risk of exceeding the range of the usize
type on any of the targets that we care about or expect the crate to be
used on. As such, using usize seems like the more general choice and
with this change we switch everything over to using usize.

+++ src/data/v2/quotes.rs
@@ -36,7 +36,7 @@ pub struct Quotes {
 #[derive(Clone, Debug, Default, PartialEq)]
 pub struct QuotesReqInit {
   /// See `QuotesReq::limit`.
-  pub limit: Option<u64>,
+  pub limit: Option<usize>,
   /// See `QuotesReq::page_token`.
   pub page_token: Option<String>,
   #[doc(hidden)]
@@ -79,7 +79,7 @@ pub struct QuotesReq {
   /// Number of quotes to return. Must be in range 1-10000, defaults to
   /// 1000.
   #[serde(rename = "limit")]
-  pub limit: Option<u64>,
+  pub limit: Option<usize>,
   /// Pagination token to continue from.
   #[serde(rename = "page_token")]
   pub page_token: Option<String>,
